Create lock design.v
Your task is to design a lock system which takes in a stream of 8-bit numbers as input and
outputs the locked/unlocked state. 
The module is locked on reset (asynchronous, active low). 
The module will look for the sequence [244, 35, 244, 6, 157] and become unlocked only if this
sequence appears at the input, while the input sequence is “valid.” 
Another input signal is used to lock the module. 
There is a “fake” sequence which is given out to suspicious people. When the sequence [88, 13,
244, 35, 90] is detected, an output indicating that a fake sequence was detected will go high.
Once this occurs, the module should refuse to unlock until a reset signal is sent.